Refactored the project structure
I also added the two system classes together
